+++ test/copyable_task.jl
@@ -4,21 +4,33 @@ using Libtask
 using Test
 
 @testset "copyable_task" begin
+    @testset "get_taped_globals outside of a task" begin
+        # This testset must come first because subsequent calls to get_taped_globals /
+        # set_taped_globals! will affect the TLS of the task running the tests.
+        @test_throws Libtask.NotInTapedTaskError Libtask.get_taped_globals(Any)
+    end
+
     for case in Libtask.TestUtils.test_cases()
         case()
     end
-    @testset "set_taped_globals!" begin
+
+    @testset "get_ and set_taped_globals!" begin
+        # Note that this testset tests both methods of get_taped_globals, the one inside the
+        # task itself, and the one outside.
         function f()
             produce(Libtask.get_taped_globals(Int))
             produce(Libtask.get_taped_globals(Int))
             return nothing
         end
         t = TapedTask(5, f)
+        @test Libtask.get_taped_globals(t) == 5
         @test consume(t) == 5
         Libtask.set_taped_globals!(t, 6)
+        @test Libtask.get_taped_globals(t) == 6
         @test consume(t) == 6
         @test consume(t) === nothing
     end
+
     @testset "iteration" begin
         function f()
             t = 1