+++ HISTORY.md
@@ -1,3 +1,14 @@
+# 0.9.12
+
+Added a new method, `Libtask.get_taped_globals(tt::TapedTask)`.
+This method extracts the taped globals stored inside `tt`.
+
+Note that this is distinct from the existing method `Libtask.get_taped_globals(::Type{T}) where {T}`.
+That method is meant for being called *inside* a TapedTask, and it extracts the taped globals from the task that it is currently inside.
+
+This PR also improves the error message for the latter method.
+When calling `Libtask.get_taped_globals(::Type{T}) where {T}` outside of a TapedTask, this now throws a `Libtask.NotInTapedTaskError` (which can be specifically caught) and shows a better error message.
+
 # 0.9.11
 
 When constructing a `TapedTask` with a method that will error when run (e.g. method doesn't exist, or is ambiguous) a more helpful error is shown.